+++ pkg/backup/table_pattern.go
@@ -61,6 +61,9 @@ func (b *Backuper) getTableListByPatternLocal(ctx context.Context, metadataPath 
 			return nil, nil, err
 		}
 	}
+	// https://github.com/Altinity/clickhouse-backup/issues/1091
+	replacer := strings.NewReplacer(`/`, "_", `\`, "_")
+
 	if err := filepath.Walk(metadataPath, func(filePath string, info os.FileInfo, err error) error {
 		if err != nil {
 			return err
@@ -79,11 +82,12 @@ func (b *Backuper) getTableListByPatternLocal(ctx context.Context, metadataPath 
 			p = strings.TrimSuffix(p, ".json")
 		}
 		names, database, table, tableName, shallSkipped, continueProcessing := b.checkShallSkipped(p, metadataPath)
-		if !continueProcessing {
+		if !continueProcessing || shallSkipped {
 			return nil
 		}
-		for _, p := range tablePatterns {
-			if matched, _ := filepath.Match(strings.Trim(p, " \t\r\n"), tableName); !matched || shallSkipped {
+		for _, pattern := range tablePatterns {
+			// https://github.com/Altinity/clickhouse-backup/issues/1091
+			if matched, _ := filepath.Match(replacer.Replace(strings.Trim(pattern, " \t\r\n")), replacer.Replace(tableName)); !matched {
 				continue
 			}
 			data, err := os.ReadFile(filePath)
@@ -232,6 +236,9 @@ func prepareTableMetadataFromSQL(data []byte, metadataPath string, names []strin
 
 func (b *Backuper) enrichTablePatternsByInnerDependencies(metadataPath string, tablePatterns []string) ([]string, error) {
 	innerTablePatterns := make([]string, 0)
+	// https://github.com/Altinity/clickhouse-backup/issues/1091
+	replacer := strings.NewReplacer("/", "_", `\`, "_")
+
 	if err := filepath.Walk(metadataPath, func(filePath string, info os.FileInfo, err error) error {
 		if err != nil {
 			return err
@@ -240,11 +247,12 @@ func (b *Backuper) enrichTablePatternsByInnerDependencies(metadataPath string, t
 			return nil
 		}
 		names, database, table, tableName, shallSkipped, continueProcessing := b.checkShallSkipped(strings.TrimSuffix(filepath.ToSlash(filePath), ".json"), metadataPath)
-		if !continueProcessing {
+		if !continueProcessing || shallSkipped {
 			return nil
 		}
-		for _, p := range tablePatterns {
-			if matched, _ := filepath.Match(strings.Trim(p, " \t\r\n"), tableName); !matched || shallSkipped {
+		for _, pattern := range tablePatterns {
+			// https://github.com/Altinity/clickhouse-backup/issues/1091
+			if matched, _ := filepath.Match(replacer.Replace(strings.Trim(pattern, " \t\r\n")), replacer.Replace(tableName)); !matched {
 				continue
 			}
 			data, err := os.ReadFile(filePath)
@@ -491,6 +499,8 @@ func getTableListByPatternRemote(ctx context.Context, b *Backuper, remoteBackupM
 		tablePatterns = strings.Split(tablePattern, ",")
 	}
 	metadataPath := path.Join(remoteBackupMetadata.BackupName, "metadata")
+	// https://github.com/Altinity/clickhouse-backup/issues/1091
+	replacer := strings.NewReplacer(`/`, "_", `\`, "_")
 	for _, t := range remoteBackupMetadata.Tables {
 		if IsInformationSchema(t.Database) {
 			continue
@@ -500,12 +510,13 @@ func getTableListByPatternRemote(ctx context.Context, b *Backuper, remoteBackupM
 			continue
 		}
 	tablePatterns:
-		for _, p := range tablePatterns {
+		for _, pattern := range tablePatterns {
 			select {
 			case <-ctx.Done():
 				return nil, ctx.Err()
 			default:
-				if matched, _ := filepath.Match(strings.Trim(p, " \t\r\n"), tableName); !matched {
+				// https://github.com/Altinity/clickhouse-backup/issues/1091
+				if matched, _ := filepath.Match(replacer.Replace(strings.Trim(pattern, " \t\r\n")), replacer.Replace(tableName)); !matched {
 					continue
 				}
 				tmReader, err := b.dst.GetFileReader(ctx, path.Join(metadataPath, common.TablePathEncode(t.Database), fmt.Sprintf("%s.json", common.TablePathEncode(t.Table))))
@@ -577,10 +588,14 @@ func parseTablePatternForDownload(tables []metadata.TableTitle, tablePattern str
 		tablePatterns = strings.Split(tablePattern, ",")
 	}
 	var result []metadata.TableTitle
+	// https://github.com/Altinity/clickhouse-backup/issues/1091
+	replacer := strings.NewReplacer("/", "_", `\`, "_")
+
 	for _, t := range tables {
 		for _, pattern := range tablePatterns {
 			tableName := fmt.Sprintf("%s.%s", t.Database, t.Table)
-			if matched, _ := filepath.Match(strings.Trim(pattern, " \t\r\n"), tableName); matched {
+			// https://github.com/Altinity/clickhouse-backup/issues/1091
+			if matched, _ := filepath.Match(replacer.Replace(strings.Trim(pattern, " \t\r\n")), replacer.Replace(tableName)); matched {
 				result = append(result, t)
 				break
 			}