+++ pkg/keeper/keeper.go
@@ -225,21 +225,23 @@ type WalkCallBack = func(node DumpNode) (bool, error)
 func (k *Keeper) Walk(prefix, relativePath string, recursive bool, callback WalkCallBack) error {
 	nodePath := path.Join(prefix, relativePath)
 	value, stat, err := k.conn.Get(nodePath)
-	log.Debug().Msgf("Walk->get(%s) = %v, err = %v", nodePath, string(value), err)
+	log.Debug().Msgf("k.Walk->get(%s) = %v, err = %v", nodePath, string(value), err)
 	if err != nil {
-		return err
+		return fmt.Errorf("k.Walk->get(%s) = %v, err = %v", nodePath, string(value), err)
 	}
 	var isDone bool
-	if isDone, err = callback(DumpNode{Path: nodePath, Value: string(value)}); err != nil {
-		return err
+	callbackNode := DumpNode{Path: nodePath, Value: string(value)}
+	if isDone, err = callback(callbackNode); err != nil {
+		return fmt.Errorf("k.Walk->callback(%v) error: %v", callbackNode, err)
 	}
 	if isDone {
 		return nil
 	}
 	if recursive && stat.NumChildren > 0 {
 		children, _, err := k.conn.Children(path.Join(prefix, relativePath))
+		log.Debug().Msgf("k.Walk->Children(%s) = %v, err = %v", path.Join(prefix, relativePath), children, err)
 		if err != nil {
-			return err
+			return fmt.Errorf("k.Walk->Children(%s) = %v, err = %v", path.Join(prefix, relativePath), children, err)
 		}
 		for _, childPath := range children {
 			if childErr := k.Walk(prefix, path.Join(relativePath, childPath), recursive, callback); childErr != nil {